Update flags in `Makefile` and compile using `-std=C99` (#37)
* Update compile flags in Makefile

And fix warnings/errors

* Replace `asprintf` with `snprintf`

* Replace `strdup` with custom `erbx_strdup`

+++ src/erbx.c
@@ -59,7 +59,7 @@ array_T* erbx_lex_file(const char* path) {
 void erbx_lex_to_buffer(char* source, buffer_T* output) {
   array_T* tokens = erbx_lex(source);
 
-  for (int i = 0; i < array_size(tokens); i++) {
+  for (size_t i = 0; i < array_size(tokens); i++) {
     token_T* token = array_get(tokens, i);
     buffer_append(output, token_to_string(token));
     buffer_append(output, "\n");
